tests/requirements/base.txt
# requirements from setup.py
djangocms-picture>=2.1.0
djangocms-link>=2.2.1
djangocms-text
setuptools

# other requirements
coverage
tox
